+++ docs/src/pages/utils/popper/popper.md
@@ -20,13 +20,6 @@ The placement of the popper updates with the available area in the viewport.
 
 {{"demo": "pages/utils/popper/SimplePopper.js" }}
 
-## PopupState helper
-
-There is a 3rd party package [`material-ui-popup-state`](https://github.com/jcoreio/material-ui-popup-state) that takes care of popper
-state for you in most cases.
-
-{{"demo": "pages/utils/popup-state/PopperPopupState.js"}}
-
 ## Scroll playground
 
 {{"demo": "pages/utils/popper/ScrollPlayground.js"}}
@@ -54,3 +47,12 @@ It is a [render props](https://reactjs.org/docs/render-props.html) demo that
 keeps track of the local state for a single popper.
 
 {{"demo": "pages/utils/popper/RenderPropsPopper.js"}}
+
+## Advanced use cases
+
+### PopupState helper
+
+There is a 3rd party package [`material-ui-popup-state`](https://github.com/jcoreio/material-ui-popup-state) that takes care of popper
+state for you in most cases.
+
+{{"demo": "pages/utils/popper/PopperPopupState.js"}}
