--- docs/src/pages/utils/popup-state/MenuPopupState.js
@@ -1,23 +1,0 @@
-import * as React from 'react';
-import Button from '@material-ui/core/Button';
-import Menu from '@material-ui/core/Menu';
-import MenuItem from '@material-ui/core/MenuItem';
-import PopupState, { bindTrigger, bindMenu } from 'material-ui-popup-state';
-
-const MenuPopupState = () => (
-  <PopupState variant="popover" popupId="demoMenu">
-    {popupState => (
-      <React.Fragment>
-        <Button variant="contained" {...bindTrigger(popupState)}>
-          Open Menu
-        </Button>
-        <Menu {...bindMenu(popupState)}>
-          <MenuItem onClick={popupState.close}>Cake</MenuItem>
-          <MenuItem onClick={popupState.close}>Death</MenuItem>
-        </Menu>
-      </React.Fragment>
-    )}
-  </PopupState>
-);
-
-export default MenuPopupState;
