--- docs/src/pages/utils/popup-state/popup-state.md
@@ -1,23 +1,0 @@
----
-title: PopupState helper component
-components: Menu, Popover, Popper
----
-
-# PopupState
-
-<p class="description">The 3rd party package <a href="https://github.com/jcoreio/material-ui-popup-state"><code>material-ui-popup-state</code></a> takes care of the boilerplate for common Menu, Popover and Popper use cases.</p>
-
-It keeps track of the local state for a single popup and passes props to inject into the
-popup component and a trigger component (like a `Button`).
-
-## Menu
-
-{{"demo": "pages/utils/popup-state/MenuPopupState.js"}}
-
-## Popover
-
-{{"demo": "pages/utils/popup-state/PopoverPopupState.js"}}
-
-## Popper
-
-{{"demo": "pages/utils/popup-state/PopperPopupState.js"}}
