+++ pages/api/menu.md
@@ -64,5 +64,4 @@ You can take advantage of this behavior to [target nested components](/guides/ap
 
 - [App Bar](/demos/app-bar/)
 - [Menus](/demos/menus/)
-- [Popup State](/utils/popup-state/)
