+++ pages/api/popover.md
@@ -74,5 +74,4 @@ You can take advantage of this behavior to [target nested components](/guides/ap
 
 - [Menus](/demos/menus/)
 - [Popover](/utils/popover/)
-- [Popup State](/utils/popup-state/)
