+++ pages/demos/menus.js
@@ -16,11 +16,11 @@ module.exports = require('fs')
   .readFileSync(require.resolve('docs/src/pages/demos/menus/SimpleMenu'), 'utf8')
 `,
         },
-        'pages/utils/popup-state/MenuPopupState.js': {
-          js: require('docs/src/pages/utils/popup-state/MenuPopupState.js').default,
+        'pages/demos/menus/MenuPopupState.js': {
+          js: require('docs/src/pages/demos/menus/MenuPopupState.js').default,
           raw: preval`
 module.exports = require('fs')
-  .readFileSync(require.resolve('docs/src/pages/utils/popup-state/MenuPopupState'), 'utf8')
+  .readFileSync(require.resolve('docs/src/pages/demos/menus/MenuPopupState'), 'utf8')
 `,
         },
         'pages/demos/menus/SimpleListMenu.js': {
